WordCamp Site Cloner: Bump temporary site limit from `50` to `125`.
A common use case is to clone the site from a previous year of the same camp (e.g., Seattle 2016 cloning Seattle 2015), but that's not possible when only 50 sites are displayed.

Show 125 sites isn't very fast or user-friendly, but it's better than preventing this use case entirely. It's also only temporary, until #1112 adds filters to improve the user experience with a large number of sites.