WordCamp.org: Create a new site a bit later.
The `wcpt_metabox_save` action is a bit misleading because
it runs multiple times, so the multi-sponsor region is saved
after the WordCamp URL is saved, causing our "create site"
functionality to fail the first time.

This commit creates a new `wcpt_metabox_save_done` action which
is called when all metabox fields are saved. Moves the site
creation process to that new action.