WordCamp: Notify when plugins have an unintended network-activation state
After a recent incident where disabling and then re-enabling WP Super Cache
caused some plugins to revert to an unwanted cached state, we realized that we
didn't have documentation anywhere of which state each plugin _should_ be in.

This MU plugin documents those intended states, and displays a notification in
the network admin if any plugin has the wrong state or is missing. This way,
intentional, permanent changes to a plugin's state will be recorded as a
committed change to this plugin.

This plugin does not force a network activation state; it only notifies when
something is amiss. The reasoning is that forcing a state has the risk of
causing weird side effects if there is ever a situation where two processes are
trying to force opposite states on the same plugin.