WordCamp QBO: Store classes in site option for reliability.
Storing it in a transient often causes the data to be empty when the QBO connection is unavailable.

This is also a good opportunity to remove some of the unnecessary "middle man" code from the `wordcamp-qbo-client` experiment, now that it's clear that the tangible drawbacks (lots of unnecessary, duplicated, unintuitive code) are worse than the intangible benefits (future-proofing for the unlikely event that we'll switch to a decentralized hosting model in the future).