WordCamp Post Type: Compare address to array item to avoid false negative.
r7629 fixed a bug where `update_venue_address()` was incorrectly returning early, but also introduced a bug where the new condition is always false, so it would never return early.

`$original_meta_values` is an array, so we need to compare `$new_address` to the first item, rather than the entire array.