WCPT: Trigger click on `Save Draft` when `Update` button is clicked in WCPT
With block editor, `Save Draft` looks like has same functionality as `Update` button used to have. We should  eventually remove the `Update` button in favor of already present `Save Draft`, but keeping it for now would make sense so as to give time to people to get used to the new editor. `Update` button is way more prominent then the `Save Draft` button, which is another reason to keep it for now.

Also linking Update to Save Draft may make it more clear that they have the same functionality.