WordCamp Post Type: Move `Event_Loader` require to plugin loader.
The class is needed for both `WordCamp_Loader` and `Meetup_Loader`, so it should be included higher up the chain, rather than relying on the assumption that `WordCamp_Loader` will load it before `Meetup_Loader` needs it.

This also makes it easier for other plugins to include `WordCamp_Loader`, in order to access it's public methods.