WordCamp Post Type: Bypass nonce verification when saving programatically.
Nonce verification was added to the `save_post` callback for WordCamp/Meetup post types in r8085. That had the side-effect of breaking the process of manually sending Organizer Reminders, because that process saves data to the corresponding WordCamp post. Because the `save_post` callback is called programatically in that scenario, there is no nonce to verify, and the `Unable to verify nonce.` error would be triggered.

We previously attempted to fix this in r8133 / r8135, but it turns out that the problem was the missing nonce rather than the post type check.