WordCamp Organizer Reminders: Only sync postmeta for WordCamp posts.
Previously this assumed that the current post being saved was a `wordcamp` post, and until recently that was always true. Something changed that allowed this to run in other contexts, like when saving an Organizer Reminder post during the process of manually sending a reminder. When that happened, the meta values for the WordCamp post would be wiped out, since there wouldn't be any matching values in `$_POST` from the Organizer Reminder post.

This is probably related to r8220, but it's not immediately obvious, and isn't worth sinking more time into.